Workaround for int-or-string format issues
Kubernetes happily expects integers to be passed to fields of type
string, if they have the format int-or-string. This means the schema is
valid (because the rules around format are hazy) but actually validating
a document against a schema will fail - correctly in my view.

The proper fix for this likely involves some discussions with api
machinery and even some spec wrangling. In the meantime I'm going to
work around it here.

This first pass is pretty blunt. It has the potential to pass some
invalid documents with a warning. With a more specific jsonpath query,
and some parsing based on the Field() it's probably possible to make
this more accurate.