This fixes the --ignore-missing-schemas flag
This changes the behavior of the --ignore-missing-schemas flag. Prior to
this change, the flag prevented validation against schemas for all
documents. The new behavior is to prevent validation against schemas for
any document whose associated schema can not be found. All other
documents are validated against their corresponding schemas.

This also improves the expected behavior of the flag. Prior behavior was
to display a successful message for any document missing its schema. New
behavior will print a warning message that the document was not
validated against a schema.