Add the --additional-schema-locations flag
This adds an option to kubeval's configuration that allows the use of
additional base URLs to search for schemas. The tool will still prefer
the default repositories, but if a schema cannot be found for a given
YAML document (e.g. a CRD), kubeval will fallback on the secondary URLs.
This implies that schemas must have been pre-generated and hosted at the
secondary URLs.