fix: TAP Schema treats SKIP as a directive, at the end of the line
Playing around with tap-to-junit vlaidations I found that the only safe way
to get the skips parsed was to use capitalized `# SKIP` with the space after
the description.

Technically lowercase seems to be allowed but the majority of the schema's
examples show capitalized.

In the schema it also says any explanation of why it was skipped should go
after the `# SKIP`, but the description of the test still goes to the left.

https://testanything.org/tap-version-13-specification.html